This addresses a few inconsistencies with our optimizer in which accidental truncation of double values to integer values via casting could lead to rewrite rules being mistakenly applied.  As an example, a rewrite rule targeting the expression 'sum(v^2)' could be mistakenly applied to the expression 'sum(v^2.3)'.

+++ src/main/java/com/ibm/bi/dml/hops/rewrite/RewriteAlgebraicSimplificationDynamic.java
@@ -1363,7 +1363,7 @@ public class RewriteAlgebraicSimplificationDynamic extends HopRewriteRule
 			//check for sum(v^2), might have been rewritten from sum(v*v)
 			if( hi2 instanceof BinaryOp && ((BinaryOp)hi2).getOp()==OpOp2.POW
 				&& hi2.getInput().get(1) instanceof LiteralOp 
-				&& HopRewriteUtils.getIntValue((LiteralOp)hi2.getInput().get(1))==2
+				&& HopRewriteUtils.getDoubleValue((LiteralOp)hi2.getInput().get(1))==2
 				&& hi2.getParent().size() == 1 ) //no other consumer than sum
 			{
 				Hop input = hi2.getInput().get(0);
